[AIRFLOW-3268] Cannot pass SSL dictionary to mysql connection via URL
It is impossible to pass 'ssl' dictionary to MySql hook as an extrai
param via URL connection. While there is a code to pass the 'ssl'
extra query parameter, MySqldb requires this parameter to be
dictionary. When you want to create a connection via URL,
you can at most have ?ssl= url-encoded string rather than
dictionary and this is how it is passed (as string). What happens
then in MySqldb, is that all SSL parameters are ignored
and MySQL establishes a non-SSL connection silently.

This is fixed by detecting if 'ssl' is a string and converting
it to json dictionary in such case.