[AIRFLOW-3524] Added support for Google Cloud Build
This commit adds support for Automated execution of Unit Tests and
(what is more important) System Tests using Google Cloud Build.

Its likely that this commit will be merged with some changes providing
that the supporting docker image that is used for Cloud Build
comes from an external repository. The ultimate goal is to
have an image that supports both Travis CI and Google Cloud Build at
the same time and it should be part of the Airflow repository itself.

Added default fernet key in postgres - same as in main test configuration